Add tests for get all parcel delivery orders [delivers #161777600] On branch ft-get-all-parcels-161777600  Changes to be committed: 	modified:   app/models/database.py 	modified:   app/views/parcels.py 	modified:   schema.sql 	modified:   test_data.sql 	modified:   tests/conftest.py 	modified:   tests/test_parcels.py 	modified:   tests/test_user.py

+++ app/models/database.py
@@ -23,7 +23,7 @@ class Database:
 
     def empty_tables(self):
        self.cursor.execute('DELETE FROM parcels')
-       self.cursor.execute('DELETE FROM users')
+       self.cursor.execute('DELETE FROM users WHERE user_id > 2')
 
 
 if __name__ == "__main__":
